+++ application/pkg/mpp/error/error.go
@@ -6,10 +6,10 @@ import (
 	"strconv"
 )
 
-func Resolve(code int) string {
+func Resolve(code int64) string {
 	switch code {
 	default:
-		out := "error " + strconv.FormatInt(int64(code), 16)
+		out := "error " + strconv.FormatInt(code, 16)
 		return out
 	}
 }
